move fetch over main
