+++ packages/insomnia/src/ui/components/toast.tsx
@@ -3,7 +3,6 @@ import type { IpcRendererEvent } from 'electron';
 import React, { FC, useEffect, useState } from 'react';
 import styled from 'styled-components';
 
-import { insomniaFetch } from '../../account/fetch';
 import * as session from '../../account/session';
 import {
   getAppId,
@@ -112,7 +111,7 @@ export const Toast: FC = () => {
         updatesNotSupported: !updatesSupported(),
         version: getAppVersion(),
       };
-      const notificationOrEmpty = await insomniaFetch<ToastNotification>({
+      const notificationOrEmpty = await window.main.insomniaFetch<ToastNotification>({
         method: 'POST',
         path: '/notification',
         obj: data,
